backend/requirements.txt
fastapi	
uvicorn	
pyjwt	
passlib[sha256_crypt]	
python-multipart	